refactor: split RPCRouterTests by subsystem
626-line monolithic file made it hard to find the test for any given
RPC method. Split into four subsystem-scoped files while keeping the
single @Suite struct + per-instance init() factory:

- RPCRouterTestHelpers.swift — the @Suite struct (db + router fixture)
- RPCRouterRepoTests.swift — repo.*, repo.updateInstructions, unknown
- RPCRouterWorktreeTests.swift — worktree.list / worktree.rename
- RPCRouterTerminalTests.swift — terminal.* including claude variant
- RPCRouterMiscTests.swift — notify, daemon.status, resolve.path, pr.*, note.*

Each per-subsystem file adds @Test methods via an extension of the
shared struct, so every test still gets a fresh router instance. No
test logic changed — same 27 tests, all green.
