+++ Tests/TBDDaemonTests/RPCRouterTerminalTests.swift
@@ -1,0 +1,119 @@
+import Testing
+import Foundation
+@testable import TBDDaemonLib
+@testable import TBDShared
+
+// Terminal-scoped RPC methods: terminal.create (shell + claude), terminal.list,
+// terminal.send, terminal.output.
+extension RPCRouterTests {
+
+    // MARK: - Terminal Tests
+
+    @Test("terminal.create and terminal.list work together")
+    func terminalCreateAndList() async throws {
+        let repo = try await db.repos.create(
+            path: "/tmp/test-repo-\(UUID().uuidString)",
+            displayName: "test-repo",
+            defaultBranch: "main"
+        )
+        let wt = try await db.worktrees.create(
+            repoID: repo.id,
+            name: "test-wt",
+            branch: "tbd/test-wt",
+            path: "/tmp/test-wt-\(UUID().uuidString)",
+            tmuxServer: "tbd-test"
+        )
+
+        // Create a terminal
+        let createReq = try RPCRequest(
+            method: RPCMethod.terminalCreate,
+            params: TerminalCreateParams(worktreeID: wt.id, cmd: "vim")
+        )
+        let createResp = await router.handle(createReq)
+        #expect(createResp.success)
+
+        let terminal = try createResp.decodeResult(Terminal.self)
+        #expect(terminal.worktreeID == wt.id)
+
+        // List terminals
+        let listReq = try RPCRequest(
+            method: RPCMethod.terminalList,
+            params: TerminalListParams(worktreeID: wt.id)
+        )
+        let listResp = await router.handle(listReq)
+        #expect(listResp.success)
+
+        let terminals = try listResp.decodeResult([Terminal].self)
+        #expect(terminals.count == 1)
+    }
+
+    @Test("terminal.send dispatches to tmux")
+    func terminalSend() async throws {
+        let repo = try await db.repos.create(
+            path: "/tmp/test-repo-\(UUID().uuidString)",
+            displayName: "test-repo",
+            defaultBranch: "main"
+        )
+        let wt = try await db.worktrees.create(
+            repoID: repo.id,
+            name: "test-wt",
+            branch: "tbd/test-wt",
+            path: "/tmp/test-wt-\(UUID().uuidString)",
+            tmuxServer: "tbd-test"
+        )
+        let terminal = try await db.terminals.create(
+            worktreeID: wt.id,
+            tmuxWindowID: "@mock-0",
+            tmuxPaneID: "%mock-0"
+        )
+
+        let request = try RPCRequest(
+            method: RPCMethod.terminalSend,
+            params: TerminalSendParams(terminalID: terminal.id, text: "echo hello")
+        )
+        let response = await router.handle(request)
+
+        // dryRun tmux should succeed
+        #expect(response.success)
+    }
+
+    // MARK: - Claude Terminal Creation
+
+    @Test("terminal.create with type claude sets label and sessionID")
+    func terminalCreateClaude() async throws {
+        let repo = try await db.repos.create(
+            path: "/tmp/test-repo-\(UUID().uuidString)",
+            displayName: "test-repo",
+            defaultBranch: "main"
+        )
+        let wt = try await db.worktrees.create(
+            repoID: repo.id,
+            name: "test-wt",
+            branch: "tbd/test-wt",
+            path: "/tmp/test-wt-\(UUID().uuidString)",
+            tmuxServer: "tbd-test"
+        )
+
+        let createReq = try RPCRequest(
+            method: RPCMethod.terminalCreate,
+            params: TerminalCreateParams(worktreeID: wt.id, type: .claude)
+        )
+        let createResp = await router.handle(createReq)
+        #expect(createResp.success)
+
+        let terminal = try createResp.decodeResult(Terminal.self)
+        #expect(terminal.label == "Claude Code")
+        #expect(terminal.claudeSessionID != nil)
+    }
+
+    // MARK: - Terminal Output
+
+    @Test("terminal.output returns error when terminal not found")
+    func terminalOutputReturnsError_whenTerminalNotFound() async throws {
+        let params = TerminalOutputParams(terminalID: UUID())
+        let request = try RPCRequest(method: RPCMethod.terminalOutput, params: params)
+        let response = await router.handle(request)
+        #expect(!response.success)
+        #expect(response.error?.contains("Terminal not found") == true)
+    }
+}
